refactor(strato): fold the Mercata Bridge into the strato adapter
Merges projects/strato-bridge into projects/strato rather than listing the bridge
as its own protocol. The inbound leg (assets escrowed in the bridge custody
address on Ethereum, Base, Linea and Robinhood Chain) becomes per-chain tvl
exports on the strato adapter, and the outbound leg (STRATO-native assets locked
in the StratoNativeBridge custody vault) is added to the strato-chain tvl.

projects/strato-bridge is removed.

Note this counts bridge custody and the STRATO-side wrapped tokens it backs under
one protocol entry: the MercataBridge assets registry maps each escrowed external
asset to its STRATO-side token (e.g. external ETH on chainId 1/8453 -> strato
93fb7295), and that token is also counted as CDP collateral, so roughly 1.42M is
represented on both legs.

Tested:
  node test.js projects/strato/index.js
  # strato 17.22M, borrowed 4.70M, ethereum 1.32M, base 76k, robinhood 23k, linea 3.1k
  # total 18.64M
